Migrate from B2 to Wasabi
B2 was a bit too slow for the website. B2 isn't designed to be a CDN,
so this is understandable. The reason that we ended up choosing Wasabi
was because we could do CNAME aliasing (cdn.wpi.moe).

We'll still be using B2 for backup, and we'll probably migrate to a
self-hosted minio instance at some point, but this is a quick fix
to get the ball rolling.